Add upgrading directory to splinter docs
In-depth guides for upgrading from one version of Splinter
to another version. Upgrade notes are only listed here if there are
changes which are important (such as breaking changes). Many versions
do not require extra work to upgrade. The release notes in each
section give an overview of all releases.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>