Move main doc index to 0.4 dir
Moves the main doc index.md to the 0.4 sub-directory, since the majority
of its content is specific to the release's documentation. The docs
directory's index file now just links to the version documentation and
the FAQ. This is an incremental change, much of this content will likely
be updated/moved in the future.

Signed-off-by: Logan Seeley <seeley@bitwise.io>