Collapse left sidebar for small screens
Collapses the left navigation sidebar on the website when the screen is
small. This is accomplished by adding a `div` to the HTML along with
some CSS rules that are based on the screen size. It also adds a button
for displaying the navigation links when the navigation is collapsed.

Signed-off-by: Logan Seeley <seeley@bitwise.io>