Update latest doc redirect to be from /docs/
Changes the redirect for the `/docs/` page to be in the `/docs/` file,
redirecting to the latest release's documentation, rather than in the
latest release, redirection from the /docs/ page. This will make it
slightly easier to update in the future; instead of moving the redirect
to the new version's file, we can just update the version number in the
existing redirect.

Signed-off-by: Logan Seeley <seeley@bitwise.io>