Add landing page for the Rust API references
This PR eliminates an external link from the left-nav TOC.

Note: The v0.5 version of the splinter and scabbard crates (and docs) aren't
being published yet, so the 0.5 doc links go to the same docs.rs location as
the 0.4 links. A later PR is needed once the crates (and docs) are versioned.

Signed-off-by: Anne Chenette <chenette@bitwise.io>