Add copyright header to docs content
Also adds a newline to the end of a few .svg logo images.

Also removes the beginning <xml> tag at the beginning of some images
in the Gameroom Walkthrough, as the presence of the license in addition
to this tag, was causing the images to not be shown on the website.

Signed-off-by: Shannyn Telander <telander@bitwise.io>