Add initial tutorial for Admin UI
This commit adds a new tutorial for administrators, "Using the Admin UI",
with copious example screenshots. Currently, the 0.4 and 0.5 versions are
identical.

This tutorial walks through the profile and circuit functions of the
Splinter Admin UI using the splinter-ui repo's docker-compose file, which
creates containers for two nodes, alpha and beta, a database container for each,
and one registry container with prepopulated key files for Alice and Bob.

To build and view the documentation:

1. docker-compose up --build

2. In a browser, go to http://127.0.0.1:8080/

3. Click on Using the Admin UI in the left-nav TOC

Signed-off-by: Anne Chenette <chenette@bitwise.io>