Update Admin UI term in Planning>Features doc
The initial planning doc used the preliminary term "Admin application" and
called it "an example of an application" (which is incorrect).

This PR updates the planning doc to use the same term as the Splinter doc,
and adds a link to the Admin UI tutorial. It also renames the file to conform
with best practices (so that file name matches the doc title).

Signed-off-by: Anne Chenette <chenette@bitwise.io>