Add splinter repository design
This change adds the initial proposal for the splinter artifact
repository layout and tools. It covers the requirements for saplings and
smart contracts, as well as introduces the concept of a Capability.

A Capability is a distribution of other artifacts that will make up a
complete application.  It defines a set of dependencies that include
smart contracts and saplings, though may include additional plugin
components, such as state delta export processors and CLI tools.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>