Add Act 2 & 3 of Gameroom walkthrough
Adds the second and third act of the Gameroom walkthrough, including the
behind the scenes sections of these acts.

Also adds formatting fixes to the tables throughout the walkthrough.
These fixes include separating large tables into separate, smaller
tables as it is in the walkthrough PDF. Also updates the text to not be
rendered as code.

Signed-off-by: Shannyn Telander <telander@bitwise.io>