Fix CLI Command Reference navigation for 0.5
Previously, the index page wasn't index.md but rather
cli_command_reference.md. This doesn't work with our approach to
automatically highlighting the section in the left navbar, because that
relies upon substring matching.

Unlike the similar change for 0.4, this change does not provide
a redirect since it is not for a stable version.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>