Update `Configuring Two Splinter Nodes`
Update the `scabbard cr create` step in the `Uploading a Smart Contract`
section of the splinter node config instructions.

Transact includes a check that ensures the scar file name matches the
contract name in the manifest, the example contract used was updated to
meet this requirement. This requires that the contract registry name
be changed to match.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>