Add splinter certificates page
This page includes in-depth information about how each certificate file is
used by splinterd as well as instructions for using Let's Encrypt certificates.

It is only included in the 0.5 docs because it focuses on a local installation
using apt and the only artifacts provided for the 0.4 branch are docker images.

Signed-off-by: Ryan Beck-Buysse <rbuysse@bitwise.io>