Add Note about required format for service args with 0.4 node
To support service-args that are not wrapped in json list,
the default formatting for the argument has changed. To be
backwards compatibly with 0.4, the service args value needs
to be explicitly set in a json list.

Also mentions --compat 0.4 for circuits compatible with 0.4.

Commit rearranges the options a little bit to make the note look
better.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>