Add a Testing document to the 0.6 release information
This includes both an overview of the LR tests and the
Grid backwards compatibility test.

The backwards compat test is the process that has been used to
verify that Grid and Splinter 0.4 and Splinter 0.6 are compatible.
This process is put up with all current hacks to be able
to do the test locally (instead of requiring docker).
In the future this hacks may not be required.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>