Add Abstract Queue design
This change adds the initial design for an abstract queue to replace the
PendingBatchQueue found in sawtooth.  It is generic over the type of
batches (or any item) it can contain as well as intended to be used in a
durable fashion.

It operates using RefTokens and TransferTokens, among other commands as
described in the RefSet feature document.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>