Update database doc - split service ID
Previously the scabbard store tables that stored service ID were storing
the FullyQualifiedServiceId in a `service_id` column. This value was
split into its two parts: circuit_id and service_id and a circuit_id
column was added to these tables. This commit Updates the splinter
database feature doc to reflect this.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>