Replace "Example Queries" section with "Views"
Replace the Example Queries section in the splinter database design doc
with a Views section that lists the current views created for scabbard
store as well as some queries to filter the views.

The new `consensus_2pc_actions_all` view accomplishes the same thing as
the example query that is being removed.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>